sync: make mutex work on unix. closes Issue #7

+++ shm/shm_linux.go
@@ -43,7 +43,7 @@ func checkShmPath(path string) bool {
 	if err := unix.Statfs(path, &statfs); err != nil {
 		return false
 	}
-	return isShmFs(statfs.Type)
+	return isShmFs(int64(statfs.Type))
 }
 
 func isShmFs(fsType int64) bool {
